pptx-service/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.32,<1
python-pptx>=1.0,<2
pydantic>=2.0,<3
pydantic-settings>=2.0,<3
python-multipart>=0.0.12
google-auth>=2.0,<3
httpx>=0.27,<1
requests>=2.28,<3
matplotlib>=3.8,<4
numpy>=1.26,<3
# P0 Features: Content Validation, Visual QA, Image Analysis
markitdown[pptx]>=0.0.2
Pillow>=10.0,<11
